stratumd: refactor vardiff and fix share difficulty consistency
- Rework VarDiff from per-share interval estimation to windowed work-rate + EMA:
  - target share time: 3s, update period: 10s, min samples: 6
  - bounded step changes: up 1.25x / down 0.8x, deadband: +/-15%
  - add desired_diff_ema and work_since_last_update state
  - reset update window on decay and no-work paths
  - extend unit tests to cover up/down/stable/capped behavior

- Fix share validation consistency with dispatched job difficulty:
  - store per-session assigned_diff (actual job difficulty sent to miner)
  - validate submitted shares against assigned_diff instead of transient worker diff
  - feed assigned_diff + network_diff into VarDiff update to avoid drift

- Normalize downstream difficulty application:
  - always dispatch with applied = min(worker_diff, network_diff)
  - refresh assigned_diff on every job broadcast
  - log worker/network/applied triplet for observability

- Harden connection behavior without breaking mining throughput:
  - exempt submit from generic request-rate disconnect path
  - keep submit throttling under share-rate limiter
  - keep stale-share accounting but avoid immediate disconnect by default
  - downgrade noisy stale upstream submit logs to debug

- Change difficulty state ownership from account-level to worker-level:
  - keep difficulty state across reconnects for the same worker login
  - prevent cross-worker contamination under the same account

- Decouple miner_client from stratumd crate internals:
  - introduce local protocol compatibility layer cmd/miner_client/src/stratum_compat.rs
  - remove direct starcoin-stratumd dependency from miner_client
  - update miner client code paths and compatibility test wiring

- Refactor stratumd entrypoint layout for maintainability:
  - split monolithic main into focused gateway/server/verify_settlement modules
  - move shared stratum constants into lib.rs for cross-module reuse
  - keep runtime behavior and protocol handling unchanged
