+++ cmd/miner_client/src/stratum_client.rs
@@ -1,6 +1,7 @@
 use crate::stratum_client_service::{
     LoginServiceRequest, ShareRequest, StratumClientService, SubmitSealRequest,
 };
+use crate::stratum_compat::{target_hex_to_difficulty, LoginRequest};
 use crate::{ConsensusStrategy, JobClient, SealEvent};
 use anyhow::Result;
 use byteorder::{LittleEndian, WriteBytesExt};
@@ -8,8 +9,6 @@ use futures::future;
 use futures::stream::{BoxStream, StreamExt};
 use starcoin_logger::prelude::error;
 use starcoin_service_registry::ServiceRef;
-use starcoin_stratumd::stratum_rpc::LoginRequest;
-use starcoin_stratumd::target_hex_to_difficulty;
 use starcoin_time_service::TimeService;
 use starcoin_types::system_events::{MintBlockEvent, MintEventExtra};
 use std::sync::Arc;
